+++ enpt/options/options_schema.py
@@ -75,6 +75,7 @@ enpt_schema_input = dict(
                 schema=dict(
                     enable_keystone_correction=dict(type='boolean', required=False),
                     enable_vnir_swir_coreg=dict(type='boolean', required=False),
+                    enable_absolute_coreg=dict(type='boolean', required=False),
                     path_reference_image=dict(type='string', required=False),
                 )),
 
@@ -149,6 +150,7 @@ parameter_mapping = dict(
     # processors > geometry
     enable_keystone_correction=('processors', 'geometry', 'enable_keystone_correction'),
     enable_vnir_swir_coreg=('processors', 'geometry', 'enable_vnir_swir_coreg'),
+    enable_absolute_coreg=('processors', 'geometry', 'enable_absolute_coreg'),
     path_reference_image=('processors', 'geometry', 'path_reference_image'),
 
     # processors > atmospheric_correction